+++ simply-synonyms/src/content/index.js
@@ -2,13 +2,15 @@ import chrome from 'browserApi'
 import googleDocsUtil from './siteLibs/googleDocsUtil'
 import api from '../api/synonyms'
 import { initializePopup, resetPopup, openPopup, getPopup, addWordsToPopup, setResultsText, stopLoading } from './popup'
+import injectPageScript, { sendPageInterfaceMessage } from './util/pageInterface'
 import './styles.css'
 
 let options = {}
 
 // Function to find selected word, fetch synonyms and open synonym popup.
 function processDoubleClick (e, w) {
-  // Processes double clicked element, or word provided
+
+  // Figure out which type of element the word is in (the target). Null means the the text isn't editable
   let targetType = null
   if (window.location.hostname === 'docs.google.com') targetType = 'gdoc'
   else if (e.target.hasAttribute('contenteditable')) targetType = 'contenteditable'
@@ -34,12 +36,19 @@ function processDoubleClick (e, w) {
     .then((response) => {
       if (response.synonyms) {
         // TODO Click callback for contenteditable and gdoc
-        const wordClickCallback = targetType !== 'input' ? null : () => {
-          e.target.value = e.target.value.slice(0, e.target.selectionStart) + word + e.target.value.slice(e.target.selectionEnd)
+        const onChooseReplacementWord = (targetType !== 'input') ? null : (wordChosen) => {
+          switch (targetType) {
+            case 'input':
+              // Replace input text with a new string containing the chosen word
+              e.target.value = e.target.value.slice(0, e.target.selectionStart) + wordChosen + e.target.value.slice(e.target.selectionEnd)
+              break
+            case 'gdoc':
+              // console.log(googleDoc.nodes)
+          }
           resetPopup()
         }
-        addWordsToPopup('synonyms', response.shortdefs, response.synonyms, wordClickCallback)
-        addWordsToPopup('antonyms', response.shortdefs, response.antonyms || [], wordClickCallback)
+        addWordsToPopup('synonyms', response.shortdefs, response.synonyms, onChooseReplacementWord)
+        addWordsToPopup('antonyms', response.shortdefs, response.antonyms || [], onChooseReplacementWord)
       } else {
         const suffixText = targetType ? 'Are you sure you spelled it correctly?' : '' // Only show this text if the target is editable
         setResultsText(`Unable to find synonyms for "${word}". ${suffixText}`)
@@ -50,9 +59,9 @@ function processDoubleClick (e, w) {
 }
 
 function addExtension() {
-  document.body.addEventListener('dblclick', processDoubleClick)
-
+  injectPageScript()
   initializePopup()
+  document.body.addEventListener('dblclick', processDoubleClick)
 }
 
 chrome.storage.local.get(['option_popupDisabled', 'option_onlyEditableText'], (result) => {